# reference/public-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "{{siteName}} Public API",
    "version": "2.0.0",
    "description": "Anonymous, read-only API for clients and Agents. No API key is required.\n\nPicking an endpoint: show recent or today's items -> /api/v1/items. Show what is trending right now -> /api/v1/hot-topics. Show the daily digest -> /api/v1/dailies/latest. Keep a complete local copy of the selected set -> /api/v1/selected/snapshot once, then /api/v1/selected/changes forever.\n\nHow often to poll: use each response's Cache-Control s-maxage as the minimum interval, and always send If-None-Match: an unchanged resource answers 304 with an empty body. On 429 or 503, honour Retry-After. There is no push channel (no SSE, webhooks or streaming): poll on an interval.\n\nTitles and summaries are generated from third-party sources; the original is always linked in links.original.",
    "termsOfService": "{{siteUrl}}/terms"
  },
  "servers": [
    {
      "url": "{{siteUrl}}"
    }
  ],
...
  "paths": {
    "/api/v1/items": {
      "get": {
        "operationId": "items",
        "summary": "List recent public AI items",
        "description": "Start here for feeds, widgets and 'what is new' views. Covers the last 7 days only. Send If-None-Match and you get a 304 with an empty body whenever nothing changed. Do not use this to maintain a complete local copy of the selected set: it is windowed, so anything older than 7 days silently falls out. Use /api/v1/selected/snapshot plus /api/v1/selected/changes for that.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          },
          {
            "name": "mode",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "selected",
                "all"
              ],
              "default": "selected"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "examples": [
                "{{categories}}"
              ],
              "description": "Current values: {{categoryList}}."
            }
          },
          {
            "name": "window",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d"
              ],
              "default": "7d"
            }
          },
          {
            "name": "by",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "timeline",
                "published"
              ],
              "default": "timeline"
            },
            "description": "Which timestamp defines the window and the ordering. 'timeline' (default) matches what the website shows: items back-filled more than 72h after publication keep their original publication date, everything else uses the time the site received it, so slow-pushing sources still land inside a 24h window. 'published' uses the third-party publication time only, which drops those slow pushes out of short windows. Both timestamps are always returned on every item as publishedAt and discoveredAt."
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 200
            },
            "description": "Leading and trailing whitespace is removed before validation; the remaining value must contain 2 to 200 Unicode code points."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Opaque, query-bound keyset cursor for the rolling window. It carries no issue time and does not expire on a timer, but the window keeps moving: once the cursor's own anchor has dropped out of the requested 24h or 7d window it is rejected with invalid_cursor, so restart the query from the first page instead of persisting a cursor across days. Older rows may also naturally leave the window between requests. Use selected snapshot + changes for an exact mirror."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemsResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hot-topics": {
      "get": {
        "operationId": "hotTopics",
        "summary": "List current multi-source hot topics",
        "description": "The current {{siteName}} Top 10, ranked by multi-source coverage and discussion heat. Each item includes a one-based rank; internal and display heat values are not returned. Uses the latest successfully published ranking, normally refreshed every five minutes. Shared caches use a 60s freshness window; poll no faster than s-maxage and use If-None-Match. If an item includes links.story, it is a human-facing HTML URL: extract its final publicId path segment and call /api/v1/stories/{publicId} for timeline JSON. Never request links.story as an API. When absent, do not construct a story id.",
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotTopicsResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ]
      }
    },
    "/api/v1/stories/{publicId}": {
      "get": {
        "operationId": "storyByPublicId",
        "summary": "Get a story (event) with its coverage timeline and AI digest",
        "description": "One story = one real-world event aggregated from every covering source: a reverse-chronological coverage timeline, an AI digest that is rewritten incrementally as the event evolves (contradictions with earlier reporting are stated explicitly), and links to same-storyline and related stories. Extract ids from the final path segment of a hot-topics links.story HTML URL, or use ids from another story's storyline/related refs; never request links.story as JSON. Merged stories answer with a 308 to the surviving id. Call this endpoint only with an id actually returned by the API; a 404 means the public story layer or that story is unavailable.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          },
          {
            "name": "publicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "308": {
            "description": "Story was merged into another; follow Location to the surviving story id."
          },
          "400": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Machine-readable problem response.",
            "headers": {
              "X-Request-Id": {
                "description": "Support correlation identifier for this failed request.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Minimum number of seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ActorUserAgent": {
        "name": "User-Agent",
        "in": "header",
        "required": false,
        "description": "The client's normal User-Agent (optional; any value works).",
        "schema": {
          "type": "string",
          "maxLength": 2048
        },
        "example": "my-reader/1.0"
      },
      "IfNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "description": "Previously received ETag for this exact URL. A match returns 304 without a response body.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "Page": {
        "type": "object",
        "required": [
          "count",
          "hasMore",
          "nextCursor"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of records in this page."
          },
          "hasMore": {
            "type": "boolean"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque; only reuse with the same query."
          }
        }
      },
      "ItemsResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "query",
          "items",
          "page"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "query": {
            "type": "object",
            "required": [
              "mode",
              "category",
              "window",
              "q",
              "by",
              "ordering"
            ],
            "properties": {
              "mode": {
                "type": "string",
                "enum": [
                  "selected",
                  "all"
                ]
              },
              "category": {
                "type": [
                  "string",
                  "null"
                ],
                "examples": [
                  "{{categories}}"
                ],
                "description": "Current values: {{categoryList}}. Clients must tolerate future string values."
              },
              "window": {
                "type": "string",
                "enum": [
                  "24h",
                  "7d"
                ]
              },
              "q": {
                "type": [
                  "string",
                  "null"
                ],
                "minLength": 2,
                "maxLength": 200
              },
              "by": {
                "type": "string",
                "enum": [
                  "timeline",
                  "published"
                ]
              },
              "ordering": {
                "type": "string",
                "enum": [
                  "timelineDesc",
                  "publishedAtDesc"
                ]
              }
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "page": {
            "$ref": "#/components/schemas/Page"
          }
        }
      },
...
      "HotTopicsResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "count",
          "items"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HotTopic"
            }
          }
        }
      },
...
      "StoryResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "story"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "story": {
            "$ref": "#/components/schemas/Story"
          }
        }
      },
...
    }
  }
}